requirements_test.txt
pytest
pytest-homeassistant-custom-component
pytest-cov
pytest-asyncio
websockets
aiohttp
voluptuous
